+++ src/actor_action_handle_berserk_transition.c
@@ -11,7 +11,7 @@
 #include "headers/blam_data_globals.h"
 
 extern void actor_berserk(uint16_t actor_index, uint8_t berserk);
-extern int actor_action_handle_combat_selection(int actor_index);
+extern uint8_t actor_action_handle_combat_selection(int actor_index);
 
 int actor_action_handle_berserk_transition(int actor_index, int16_t berserk_level)
 {
