+++ src/actor_type_get_when_to_pursue.c
@@ -1,7 +1,7 @@
 #include <stdint.h>
 #include "headers/actor_type_definition.h"
 
-int actor_type_get_when_to_pursue(int16_t actor_type)
+int16_t actor_type_get_when_to_pursue(int16_t actor_type)
 {
-    return (uint16_t)actor_type_definitions[actor_type]->when_to_pursue;
+    return actor_type_definitions[actor_type]->when_to_pursue;
 }
