+++ src/ai_scripting_status_evaluate.c
@@ -5,7 +5,7 @@
 
 extern int * hs_macro_function_evaluate(int16_t function_index, int thread_index, uint8_t initialize);
 extern void hs_return(uint16_t thread_index, int value);
-extern int ai_scripting_status(int ai_index);
+extern int16_t ai_scripting_status(int ai_index);
 
 void ai_scripting_status_evaluate(int16_t function_index, int thread_index, uint8_t initialize)
 {
