+++ src/local_random_range.c
@@ -6,7 +6,7 @@
 extern uint32_t *get_global_local_random_seed_address(void);
 extern int16_t seed_random_range(uint32_t *seed, int16_t lower_bound, int16_t upper_bound);
 
-int local_random_range(int16_t lower_bound, int16_t upper_bound)
+int16_t local_random_range(int16_t lower_bound, int16_t upper_bound)
 {
     return seed_random_range(get_global_local_random_seed_address(), lower_bound, upper_bound);
 }
