+++ src/recorded_animation_get_time_left.c
@@ -1,13 +1,14 @@
-/* recorded_animation_get_time_left @0x837125A0 — ticks remaining on the recorded animation playing on a
+/* recorded_animation_get_time_left @0x83712578 — ticks remaining on the recorded animation playing on a
  * unit, or 0 if none (animation_thread.unit_index / .ticks_left). */
 
+#include <stdint.h>
 #include "headers/animation_thread.h"
 #include "headers/data_iterator.h"
 #include "headers/data_array.h"
 #include "headers/blam_data_globals.h"
 
 
-int recorded_animation_get_time_left(int unit_index)
+uint16_t recorded_animation_get_time_left(int unit_index)
 {
     data_iterator iterator;
     data_iterator_new(&iterator, animation_threads);
