+++ src/scenario_test_pas.c
@@ -10,7 +10,7 @@
 extern unsigned int *structure_bsp_get_cluster_pvs(structure_bsp *structure_bsp, int16_t cluster_index);
 extern uint8_t bit_vector_and(int16_t count, const unsigned int *a, const unsigned int *b, unsigned int *result);
 
-int scenario_test_pas(int16_t cluster_index0, int16_t cluster_index1)
+uint8_t scenario_test_pas(int16_t cluster_index0, int16_t cluster_index1)
 {
     const unsigned int *pvs0 = structure_bsp_get_cluster_pvs(global_structure_bsp, cluster_index0);
     const unsigned int *pvs1 = structure_bsp_get_cluster_pvs(global_structure_bsp, cluster_index1);
