+++ src/scenery_get_animation_time.c
@@ -1,10 +1,11 @@
-/* scenery_get_animation_time @0x837EA008 — ticks remaining (minus a 2-frame lead) on a scenery object's
+/* scenery_get_animation_time @0x837E9FE8 — ticks remaining (minus a 2-frame lead) on a scenery object's
  * current animation, or 0 if it isn't animating (scenery flag bit 0 clear). The current animation's frame
  * count is read from the model-animation tag (object animation graph index, animations block, animation
  * element frame_count); the current frame is the object's animation frame index.
  *
  * DEVIATION: the decompiler's sign-bit/equality bit-twiddle reduces to max(frames_left - 2, 0). */
 
+#include <stdint.h>
 #include "headers/data_array.h"
 #include "headers/global_tag_instances.h"
 #include "headers/object_header_datum.h"
@@ -15,7 +16,7 @@
 #include "headers/blam_data_globals.h"
 
 
-int scenery_get_animation_time(int scenery_index)
+int16_t scenery_get_animation_time(int scenery_index)
 {
     scenery_datum *scenery = (scenery_datum *)DATA_ARRAY_ELEMENT(object_header_data, object_header_datum, scenery_index)->datum;
     if ( (scenery->scenery.flags & (1u << _scenery_self_animated_bit)) == 0 )
@@ -27,5 +28,5 @@ int scenery_get_animation_time(int scenery_index)
     short current_frame = scenery->object.animation.state.frame_index;
 
     int frames_left = total_frames - current_frame;
-    return frames_left > 2 ? frames_left - 2 : 0;
+    return (int16_t)(frames_left > 2 ? frames_left - 2 : 0);   /* extsh r3,r5 @0x837EA070 */
 }
